requirements.txt
# Core Scientific Computing
numpy==2.4.2
pandas==2.3.3
scipy==1.17.1

# Machine Learning
scikit-learn==1.8.0

# Materials Science
mp-api==0.46.0
pymatgen==2026.3.23

# Visualization
matplotlib==3.10.8
seaborn==0.13.2

# Utilities
pyyaml==6.0.3
python-dotenv==1.2.2

# Testing (optional, for development)
# pytest>=7.0.0
# pytest-cov>=3.0.0
